+++ concordium-sdk-examples/src/main/java/com/concordium/sdk/examples/GetAccountInfo.java
@@ -16,7 +16,7 @@ import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.concurrent.Callable;
 
-@Command(name = "GetAnonymityRevokers", mixinStandardHelpOptions = true)
+@Command(name = "GetAccountInfo", mixinStandardHelpOptions = true)
 public class GetAccountInfo implements Callable<Integer> {
     @Option(
             names = {"--endpoint"},